requirements.txt
aiofiles==24.1.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask_Cors==5.0.0
numpy==2.2.3
pdf2image==1.17.0
pdfplumber==0.11.4
Pillow==11.1.0
pytesseract==0.3.13
python-dotenv==1.0.1
Requests==2.32.3
torch==2.5.1+cpu
transformers==4.48.0
